backend/requirements.txt
fastapi
uvicorn
google-generativeai
onnxruntime
tokenizers
numpy
huggingface-hub
pdfplumber
python-multipart
requests
python-dotenv
gunicorn
